Initial Commit
This is a fork of the pydynamixel project from: http://code.google.com/p/pydynamixel/

Original Description from that page:

This is a Python version of the ForestMoon Dynamixel library originally written in C# by Scott Ferguson. The Python version was created for the Pi Robot Project (mailto:patrick@pirobot.org) which lives at http://www.pirobot.org.

The original license for the C# version is as follows:

This software was written and developed by Scott Ferguson. The current version can be found at http://www.forestmoon.com/Software/. This free software is distributed under the GNU General Public License. See http://www.gnu.org/licenses/gpl.html for details. This license restricts your usage of the software in derivative works.

Here are the instructions for installing the library:

Get the installer file from the Downloads tab above.
Unzip the archive to a location outside of your Python installation.
Bring up a command prompt and cd into the dynamixel-installer directory.
Run the command "python setup.py install"
You can now use the library by doing an "import dynamixel".
For an example of how to use it, see the two sample files example.py and example2.py in the dynamixel-installer directory.